+++ crates/client/exec/src/block_context.rs
@@ -15,7 +15,8 @@ use std::sync::Arc;
 pub struct ExecutionContext {
     pub(crate) backend: Arc<MadaraBackend>,
     pub(crate) block_context: BlockContext,
-    pub(crate) db_id: DbBlockId,
+    /// None means we are executing the genesis block. (no latest block)
+    pub(crate) latest_visible_block: Option<DbBlockId>,
 }
 
 impl ExecutionContext {
@@ -33,53 +34,91 @@ impl ExecutionContext {
     }
 
     pub fn init_cached_state(&self) -> CachedState<BlockifierStateAdapter> {
-        let on_top_of = match self.db_id {
-            DbBlockId::Pending => Some(DbBlockId::Pending),
-            DbBlockId::Number(block_n) => {
-                // We exec on top of the previous block. None means we are executing genesis.
-                block_n.checked_sub(1).map(DbBlockId::Number)
-            }
-        };
-
         tracing::debug!(
             "Init cached state on top of {:?}, block number {:?}",
-            on_top_of,
+            self.latest_visible_block,
             self.block_context.block_info().block_number.0
         );
 
         CachedState::new(BlockifierStateAdapter::new(
             Arc::clone(&self.backend),
             self.block_context.block_info().block_number.0,
-            on_top_of,
+            self.latest_visible_block,
         ))
     }
 
-    /// Create an execution context for executing transactions **within** that block.
+    /// Init execution at the beginning of a block. The header of the block will be used, but all of the
+    /// transactions' state modifications will not be visible.
+    ///
+    /// This function is usually what you would want for the `trace` rpc enpoints, for example.
     #[tracing::instrument(skip(backend, block_info), fields(module = "ExecutionContext"))]
-    pub fn new_in_block(backend: Arc<MadaraBackend>, block_info: &MadaraMaybePendingBlockInfo) -> Result<Self, Error> {
-        let (db_id, protocol_version, block_number, block_timestamp, sequencer_address, l1_gas_price, l1_da_mode) =
-            match block_info {
-                MadaraMaybePendingBlockInfo::Pending(block) => (
-                    DbBlockId::Pending,
-                    block.header.protocol_version,
-                    // when the block is pending, we use the latest block n + 1
+    pub fn new_at_block_start(
+        backend: Arc<MadaraBackend>,
+        block_info: &MadaraMaybePendingBlockInfo,
+    ) -> Result<Self, Error> {
+        let (latest_visible_block, header_block_id) = match block_info {
+            MadaraMaybePendingBlockInfo::Pending(_block) => {
+                let latest_block_n = backend.get_latest_block_n()?;
+                (
+                    latest_block_n.map(DbBlockId::Number),
+                    // when the block is pending, we use the latest block n + 1 to make the block header
                     // if there is no latest block, the pending block is actually the genesis block
-                    backend.get_latest_block_n()?.map(|el| el + 1).unwrap_or(0),
-                    block.header.block_timestamp,
-                    block.header.sequencer_address,
-                    block.header.l1_gas_price.clone(),
-                    block.header.l1_da_mode,
-                ),
-                MadaraMaybePendingBlockInfo::NotPending(block) => (
-                    DbBlockId::Number(block.header.block_number),
-                    block.header.protocol_version,
-                    block.header.block_number,
-                    block.header.block_timestamp,
-                    block.header.sequencer_address,
-                    block.header.l1_gas_price.clone(),
-                    block.header.l1_da_mode,
-                ),
-            };
+                    latest_block_n.map(|el| el + 1).unwrap_or(0),
+                )
+            }
+            MadaraMaybePendingBlockInfo::NotPending(block) => {
+                // If the block is genesis, latest visible block is None.
+                (block.header.block_number.checked_sub(1).map(DbBlockId::Number), block.header.block_number)
+            }
+        };
+        Self::new(backend, block_info, latest_visible_block, header_block_id)
+    }
+
+    /// Init execution on top of a block. All of the transactions' state modifications are visible
+    /// but the execution still happens within that block.
+    /// This is essentially as if we're executing on top of the block after all of the transactions
+    /// are executed, but before we switched to making a new block.
+    ///
+    /// This function is usually what you would want for the `estimateFee`, `simulateTransaction`, `call` rpc endpoints, for example.
+    #[tracing::instrument(skip(backend, block_info), fields(module = "ExecutionContext"))]
+    pub fn new_at_block_end(
+        backend: Arc<MadaraBackend>,
+        block_info: &MadaraMaybePendingBlockInfo,
+    ) -> Result<Self, Error> {
+        let (latest_visible_block, header_block_id) = match block_info {
+            MadaraMaybePendingBlockInfo::Pending(_block) => {
+                let latest_block_n = backend.get_latest_block_n()?;
+                (Some(DbBlockId::Pending), latest_block_n.map(|el| el + 1).unwrap_or(0))
+            }
+            MadaraMaybePendingBlockInfo::NotPending(block) => {
+                (Some(DbBlockId::Number(block.header.block_number)), block.header.block_number)
+            }
+        };
+        Self::new(backend, block_info, latest_visible_block, header_block_id)
+    }
+
+    fn new(
+        backend: Arc<MadaraBackend>,
+        block_info: &MadaraMaybePendingBlockInfo,
+        latest_visible_block: Option<DbBlockId>,
+        block_number: u64,
+    ) -> Result<Self, Error> {
+        let (protocol_version, block_timestamp, sequencer_address, l1_gas_price, l1_da_mode) = match block_info {
+            MadaraMaybePendingBlockInfo::Pending(block) => (
+                block.header.protocol_version,
+                block.header.block_timestamp,
+                block.header.sequencer_address,
+                block.header.l1_gas_price.clone(),
+                block.header.l1_da_mode,
+            ),
+            MadaraMaybePendingBlockInfo::NotPending(block) => (
+                block.header.protocol_version,
+                block.header.block_timestamp,
+                block.header.sequencer_address,
+                block.header.l1_gas_price.clone(),
+                block.header.l1_da_mode,
+            ),
+        };
 
         let versioned_constants = backend.chain_config().exec_constants_by_protocol_version(protocol_version)?;
         let chain_info = ChainInfo {
@@ -106,7 +145,7 @@ impl ExecutionContext {
                 versioned_constants,
                 backend.chain_config().bouncer_config.clone(),
             ),
-            db_id,
+            latest_visible_block,
             backend,
         })
     }